% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_barcharts.R
\name{generate_bar_charts}
\alias{generate_bar_charts}
\title{a function to generate bar charts for categorical data}
\usage{
generate_bar_charts(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
plot
}
\description{
a function to generate bar charts for categorical data
}
