% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_plots.R
\name{generate_boxplots}
\alias{generate_boxplots}
\title{a function to generate boxplots and density plots for continuous data}
\usage{
generate_boxplots(data, save_plots = FALSE, output_dir = "boxplots")
}
\arguments{
\item{data}{a dataframe}

\item{save_plots}{boolean, if TRUE will save plots in specified directory}

\item{output_dir}{directory for saving plots}
}
\value{
p
}
\description{
a function to generate boxplots and density plots for continuous data
}
