% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_tables.R
\name{summarize_cat}
\alias{summarize_cat}
\title{a function to generate a summary table for categorical data}
\usage{
summarize_cat(
  data,
  percentage_missing = 30,
  SMD_flag = 0.2,
  cat_raw_output = FALSE
)
}
\arguments{
\item{data}{a dataframe}

\item{percentage_missing}{numeric flag for percent missing data}

\item{SMD_flag}{numeric flag for standardized mean difference}

\item{cat_raw_output}{boolean indicator, TRUE will return raw output table and FALSE will return kable table}
}
\value{
kable_table
}
\description{
a function to generate a summary table for categorical data
}
