% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_tables.R
\name{summarize_cont}
\alias{summarize_cont}
\title{a function to generate a summary table for continuous variables}
\usage{
summarize_cont(
  data,
  cv_flag = 30,
  missing_flag = 5,
  skewness_flag = 2,
  kurtosis_flag = 2,
  outlier_flag = 5,
  cont_raw_output = FALSE
)
}
\arguments{
\item{data}{a dataframe}

\item{cv_flag}{numeric flag for coefficient of variation}

\item{missing_flag}{numeric flag for percent of missing data}

\item{skewness_flag}{numeric flag for skewness level}

\item{kurtosis_flag}{numeric flag for kurtosis level}

\item{outlier_flag}{numeric flag for number of outliers}

\item{cont_raw_output}{boolean, if TRUE, will return raw output, if FALSE, will return kable table}
}
\value{
kable_table
}
\description{
a function to generate a summary table for continuous variables
}
