% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniEDA.R
\name{uniEDA}
\alias{uniEDA}
\title{univariate exploratory data analysis}
\usage{
uniEDA(
  data,
  cv_flag = 30,
  missing_flag = 5,
  skewness_flag = 2,
  kurtosis_flag = 2,
  outlier_flag = 5,
  min_category = 3,
  percentage_flag = 30,
  SMD_flag = 0.2,
  cont_boxplots = FALSE,
  cont_densplots = FALSE,
  cat_barcharts = FALSE,
  cont_raw_output = FALSE,
  cat_raw_output = FALSE,
  exclud_vars = NULL
)
}
\arguments{
\item{data}{data frame, any dataset containing continuous and/or categorical data}

\item{cv_flag}{numeric, coefficient of variation flag in continuous table, calculated as SD/mean*100, default set to 30}

\item{missing_flag}{numeric, flag for percentage of missing data in continuous table, default set to 5%}

\item{skewness_flag}{numeric, flag for skewness in continuous table, default set to 2}

\item{kurtosis_flag}{numeric, flag for excess kurtosis (kurtosis - 3) in continuous table, default set to 2}

\item{outlier_flag}{numeric, flag for percentage of data that are outliers in continuous table, default set to 5%}

\item{min_category}{numeric, minimum number of categories to be considered as categorical variables, default set to 3}

\item{percentage_flag}{numeric, flag for missing percent in categorical table, default set to 30}

\item{SMD_flag}{numeric, flag for standardized mean difference in categorical table, default set to .2}

\item{cont_boxplots}{logical, include continuous box plots TRUE or FALSE}

\item{cont_densplots}{logical, include continuous density plots TRUE or FALSE}

\item{cat_barcharts}{logical, include categorical bar charts TRUE or FALSE}

\item{cont_raw_output}{logical, if TRUE, returns continuous summary table as raw data, if FALSE, returns kable table}

\item{exclud_vars}{character, variables in dataset to be excluded}
}
\value{
Continuous summary table: A tibble (if `cont_raw_output = TRUE`) or a `kableExtra` object (if `cont_raw_output = FALSE`):

- **Variable**: Name of the variable.
- **N**: Number of non-missing values.
- **Missing**: Number of missing values.
- **Missingpercent**: Percentage of missing values.
- **Mean**: Mean of the variable.
- **Median**: Median of the variable.
- **SD**: Standard deviation.
- **Min**: Minimum value.
- **Max**: Maximum value.
- **Q1**: First quartile.
- **Q3**: Third quartile.
- **IQR**: Interquartile range.
- **Skewness**: Skewness of the variable.
- **Kurtosis**: Kurtosis of the variable.
- **NOutliers**: Number of outliers.
- **PercentOutliers**: Percentage of outliers.
- **Flags**: Columns for flagged variables, such as `SkewnessFlag`, `KurtosisFlag`, `MissingFlag`, `CVFlag`, and `OutlierFlag`.

Categorical summary table: A tibble (if `cat_raw_output = TRUE`) or a `kableExtra` object (if `cat_raw_output = FALSE`):

   - **Variable**: Name of the variable.
   - **Level**: Category levels.
   - **Frequency**: Frequency of each category.
   - **Percent**: Frequency percentage of each category.
   - **SMD**: Standardized mean difference between levels

Boxplots for Continuous Variables (if cont_boxplots = TRUE)
   A `ggplot2` object showing boxplots of continuous variables.

Density Plots for Continuous Variables (if cont_densplots = TRUE)
   A `ggplot2` object showing density plots for continuous variables.

Bar Charts for Categorical Variables (if cat_barcharts = TRUE)
   A `ggplot2` object showing bar charts for categorical variables
}
\description{
a function that returns descriptive tables and figures for continuous and categorical data
}
\examples{
uniEDA(armed_conf, cont_boxplots = TRUE, cont_densplots = TRUE, cat_barcharts = TRUE, exclud_vars = "ISO")

Example data from
Jawad M, Hone T, Vamos EP, Cetorelli V, Millett C.
Implications of armed conflict for maternal and child health: A regression analysis of data from 181 countries for 2000-2019.
PLoS Med. 2021 Sep 28;18(9):e1003810. doi: 10.1371/journal.pmed.1003810. PMID: 34582455; PMCID: PMC8478221.

}
\author{
Meagan Lacroix, Rebecca Raj, Syeda Aiman Fatima, Xinze Yu, Xingchen Hu
}
